change_zulip_topic: Don't send automated notifications
The change is being performed, per message. Sending a notification for
each message can get very noisy. We could look at sending a notification
on the last change, but also the script might be run manually a bunch of
times, in which case the notifications can be annoying. May be we just
use a flag in the commandline for notifications, if we need to.